Try to understand ownership in Rust
This is my second time reading about ownership and it makes much more sense to me this time! Probably because I learned more CS along the way. But I still don't fully understand the reasoning behind the reference rules. I understand it intuitively, but I'd like to understand it in a math-proof-like way.